requirements.txt
streamlit
yfinance
pandas
plotly
matplotlib
bcrypt
python-dotenv
